requirements.txt
# requirements.txt
# Install all dependencies with:
#   pip install -r requirements.txt

torch
torchvision
matplotlib
Pillow
numpy